requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-mysql==3.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Faker==8.1.1
mysqlclient==2.0.3
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
